requirements.txt
# Web 框架
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1

# 数据库
pymysql==1.1.1

# 认证
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4

# HTTP 客户端
httpx==0.28.1

# 原有依赖 (Dify 模块和调度器仍在使用)
requests==2.32.3
urllib3==2.4.0
